+++ auto-instrumentation/okhttp/okhttp-3.0/library/src/main/java/io/opentelemetry/instrumentation/library/okhttp/v3_0/internal/OkHttp3Singletons.java
@@ -11,7 +11,7 @@ import io.opentelemetry.api.GlobalOpenTelemetry;
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
 import io.opentelemetry.instrumentation.api.instrumenter.Instrumenter;
-import io.opentelemetry.instrumentation.api.instrumenter.http.HttpClientResend;
+import io.opentelemetry.instrumentation.api.instrumenter.http.HttpClientResendCount;
 import io.opentelemetry.instrumentation.api.instrumenter.net.PeerServiceAttributesExtractor;
 import io.opentelemetry.instrumentation.library.okhttp.v3_0.OkHttpInstrumentationConfig;
 import io.opentelemetry.instrumentation.okhttp.v3_0.internal.ConnectionErrorSpanInterceptor;
@@ -38,15 +38,18 @@ public final class OkHttp3Singletons {
                                             OkHttpInstrumentationConfig
                                                     .getCapturedResponseHeaders())
                                     .setKnownMethods(OkHttpInstrumentationConfig.getKnownMethods()),
+                    spanNameExtractorConfigurer -> {
+                        //nop?
+                    },
                     singletonList(
-                            PeerServiceAttributesExtractor.create(
-                                    OkHttpAttributesGetter.INSTANCE,
-                                    OkHttpInstrumentationConfig.getPeerServiceMapping())),
+                                    PeerServiceAttributesExtractor.create(
+                                            OkHttpAttributesGetter.INSTANCE,
+                                            OkHttpInstrumentationConfig.newPeerServiceResolver())),
                     OkHttpInstrumentationConfig.emitExperimentalHttpClientMetrics());
 
     public static final Interceptor CONTEXT_INTERCEPTOR =
             chain -> {
-                try (Scope ignored = HttpClientResend.initialize(Context.current()).makeCurrent()) {
+                try (Scope ignored = HttpClientResendCount.initialize(Context.current()).makeCurrent()) {
                     return chain.proceed(chain.request());
                 }
             };