+++ instrumentation/src/test/java/io/opentelemetry/android/instrumentation/crash/CrashReporterTest.java
@@ -18,7 +18,7 @@ import io.opentelemetry.sdk.OpenTelemetrySdk;
 import io.opentelemetry.sdk.testing.assertj.TraceAssert;
 import io.opentelemetry.sdk.testing.junit5.OpenTelemetryExtension;
 import io.opentelemetry.sdk.trace.data.StatusData;
-import io.opentelemetry.semconv.trace.attributes.SemanticAttributes;
+import io.opentelemetry.semconv.SemanticAttributes;
 import java.time.Duration;
 import java.util.function.Consumer;
 import org.junit.jupiter.api.AfterAll;